rename command execution files to <command>-command.ts to avoid duplicate names
  + regularize source code a bit
