+++ packages/cli/src/m3/diagram-command.ts
