+++ packages/cli/src/serialization/diff-command.ts
