+++ packages/cli/src/serialization/textualize-command.ts
